Refactor create_srg_export again
